# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/field/:fieldId/variants": {
            "get": {
                "description": "Get Field Variants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldVariants"
                ],
                "summary": "GetFieldVariants",
                "operationId": "get-field-variants",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FieldVariants"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Field Variants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldVariants"
                ],
                "summary": "CreateFieldVariants",
                "operationId": "create-field-variants",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FieldVariants"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FieldVariants"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Field Variants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldVariants"
                ],
                "summary": "DeleteFieldVariants",
                "operationId": "delete-field-variants",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Field Variant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FieldVariants"
                ],
                "summary": "UpdateFieldVariants",
                "operationId": "update-field-variants",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FieldVariants"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FieldVariants"
                        }
                    }
                }
            }
        },
        "/api/v1/form": {
            "get": {
                "description": "Get All Forms",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Form"
                ],
                "summary": "GetAllForms",
                "operationId": "get-all-forms",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Form"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Form"
                ],
                "summary": "CreateForm",
                "operationId": "create-form",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Form"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Form"
                        }
                    }
                }
            }
        },
        "/api/v1/form/:formId/field": {
            "get": {
                "description": "Get All Fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Field"
                ],
                "summary": "GetAllFields",
                "operationId": "get-all-fields",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Field"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Field"
                ],
                "summary": "CreateField",
                "operationId": "create-field",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Field"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Field"
                        }
                    }
                }
            }
        },
        "/api/v1/form/:formId/field/:id": {
            "get": {
                "description": "Get One Field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Field"
                ],
                "summary": "GetOneField",
                "operationId": "get-one-field",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Field"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Field"
                ],
                "summary": "DeleteField",
                "operationId": "delete-field",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Field"
                ],
                "summary": "UpdateField",
                "operationId": "update-field",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Field"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Field"
                        }
                    }
                }
            }
        },
        "/api/v1/form/:id": {
            "get": {
                "description": "Get One Form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Form"
                ],
                "summary": "GetOneForm",
                "operationId": "get-one-form",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Form"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Form"
                ],
                "summary": "DeleteForm",
                "operationId": "delete-form",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Form"
                ],
                "summary": "UpdateForm",
                "operationId": "update-form",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Form"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Form"
                        }
                    }
                }
            }
        },
        "/api/v1/permission": {
            "get": {
                "description": "Get All Permissions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "GetAllPermissions",
                "operationId": "get-all-permissions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.UserPermission"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Permission",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "CreatePermission",
                "operationId": "create-permission",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserPermission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.UserPermission"
                        }
                    }
                }
            }
        },
        "/api/v1/permission/:id": {
            "get": {
                "description": "Get One Permission",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "GetOnePermission",
                "operationId": "get-one-permission",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.UserPermission"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Permission",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "DeletePermission",
                "operationId": "delete-permission",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Permission",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "UpdatePermission",
                "operationId": "update-permission",
                "parameters": [
                    {
                        "description": "body info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserPermission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.UserPermission"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.Field": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order_of": {
                    "type": "integer"
                },
                "required": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "request.FieldVariants": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "variant": {
                    "type": "string"
                }
            }
        },
        "request.Form": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "integer"
                },
                "logo": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "request.UserPermission": {
            "type": "object",
            "properties": {
                "permission": {
                    "type": "string"
                }
            }
        },
        "response.Field": {
            "type": "object",
            "properties": {
                "form_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order_of": {
                    "type": "integer"
                },
                "required": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "response.FieldVariants": {
            "type": "object",
            "properties": {
                "field_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "variant": {
                    "type": "string"
                }
            }
        },
        "response.Form": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "response.UserPermission": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "permission": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}